Add query service
