+++ src/main/java/org/bih/aft/ports/QueryUseCase.java
@@ -1,7 +1,7 @@
 package org.bih.aft.ports;
 
 import org.bih.aft.controller.dao.AQLinput;
-import org.bih.aft.service.dao.FeasabilityOutput;
+import org.bih.aft.service.dao.FeasibilityOutput;
 
 import java.util.List;
 
@@ -9,7 +9,7 @@ import java.util.List;
  * Input port for querying.
  */
 public interface QueryUseCase {
-    List<FeasabilityOutput> federate(AQLinput query);
+    List<FeasibilityOutput> federate(AQLinput query);
 
-    FeasabilityOutput local(AQLinput query);
+    FeasibilityOutput local(AQLinput query);
 }
